ci,docs,toolbox: re-vamp dependencies [wip]
Generally, then the intent is to verify compatibility with the upcoming
release, latest stable/lts, and the previous stable/lts. This is for the
most popular Linux distributions three different versions, except for
rolling distros such as Arch Linux, for these only the "latest" version
is checked.

Experimenting
=============

The ability too install Python packages system-wide is being removed
from Linux distributions. One work-around for this is using pipx for
command-line tools. This is experimented with here, as an ability to
install via pipx.

However, for libraries another solution is needed which is better than
'--break-system-packages'.

Changes to pkgs-scripts
=======================

* Add list of packages "inline" instead using .txt file
* All shell-shebangs are now "!/bin/sh"
* Changed installation order to always install
  * Via system package manager
  * Via source-archive
  * Via pip

Specific changes follow.

Alpine
------

* Add ld-command

Debian
------

* bookworm/testing: install liburing via package-repository

Ubuntu
------

* Add Kinetic Kudu, as a means to check the latest Ubuntu release
* Drop Bionic as it is going EOL soon (April 2023)

Fedora
------

Checking compatibility with Fedora is intented to verify the
upcoming/rawhide release, the current stable, and the previous stable.

* Add Fedora 38, 37
* Change Fedora 36 to install liburing from source instead of from
  rawhide
* Drop Fedora 35 and 34 as they are EOL

Windows #wip
------------

The scripts are quite simply moved into the template for Windows. It
does not behave as it is supposed to.

CI
==

The GitHUB Action build-linux is altered to run Fedora 38, 37, and
dropped building on Fedora 35, 34 as they are EOL.

TODO / Not-checked
==================

* Generate docs/getting_stated/toolchain based on deps.yaml
* Add docs for kinetic
* Add kinetic to CI

Signed-off-by: Simon A. F. Lund <os@safl.dk>

ci,docs,toolbox: re-vamp dependencies [wip]

Signed-off-by: Simon A. F. Lund <os@safl.dk>